requirement.txt
pandas
pandas_ta
yfinance
python-dotenv
python-telegram-bot
requests
pytz
lxml